+++ packages/devtools_app_shared/test/ui/ide_theme_test.dart
@@ -4,6 +4,7 @@
 
 import 'dart:ui';
 
+import 'package:devtools_app_shared/shared.dart';
 import 'package:devtools_app_shared/ui.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -11,7 +12,7 @@ void main() {
   group('$IdeThemeQueryParams', () {
     test('successfully creates params', () {
       final params = IdeThemeQueryParams({
-        'embed': 'true',
+        'embedMode': 'one',
         'backgroundColor': '#112233',
         'foregroundColor': '#112244',
         'fontSize': '8.0',
@@ -19,7 +20,7 @@ void main() {
       });
 
       expect(params.params, isNotEmpty);
-      expect(params.embed, true);
+      expect(params.embedMode, EmbedMode.embedOne);
       expect(params.backgroundColor, const Color(0xFF112233));
       expect(params.foregroundColor, const Color(0xFF112244));
       expect(params.fontSize, 8.0);
@@ -28,7 +29,7 @@ void main() {
 
     test('handles bad input', () {
       final params = IdeThemeQueryParams({
-        'embed': 'true',
+        'embedMode': 'blah',
         'backgroundColor': 'badcolor',
         'foregroundColor': 'badcolor',
         'fontSize': 'eight',
@@ -36,7 +37,7 @@ void main() {
       });
 
       expect(params.params, isNotEmpty);
-      expect(params.embed, true);
+      expect(params.embedMode, EmbedMode.none);
       expect(params.backgroundColor, isNull);
       expect(params.foregroundColor, isNull);
       expect(params.fontSize, unscaledDefaultFontSize);
@@ -46,7 +47,7 @@ void main() {
     test('creates empty params', () {
       final params = IdeThemeQueryParams({});
       expect(params.params, isEmpty);
-      expect(params.embed, false);
+      expect(params.embedMode, EmbedMode.none);
       expect(params.backgroundColor, isNull);
       expect(params.foregroundColor, isNull);
       expect(params.fontSize, unscaledDefaultFontSize);